moved grid type menu
