+++ packages/tldraw/src/components/TopPanel/PageMenu/PageMenu.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react'
 import * as DropdownMenu from '@radix-ui/react-dropdown-menu'
-import { PlusIcon, CheckIcon } from '@radix-ui/react-icons'
+import * as ToggleGroup from '@radix-ui/react-toggle-group'
+import { PlusIcon, CheckIcon, Cross2Icon, DragHandleDots2Icon, FrameIcon, TextAlignJustifyIcon, TriangleRightIcon, SpeakerLoudIcon } from '@radix-ui/react-icons'
 import { PageOptionsDialog } from '../PageOptionsDialog'
 import { styled } from '~styles'
 import { useTldrawApp } from '~hooks'
@@ -14,9 +15,7 @@ import { FormattedMessage } from 'react-intl'
 const sortedSelector = (s: TDSnapshot) =>
   Object.values(s.document.pages).sort((a, b) => (a.childIndex || 0) - (b.childIndex || 0))
 
-const currentPageNameSelector = (s: TDSnapshot) => s.document.pages[s.appState.currentPageId].name
-
-const currentPageIdSelector = (s: TDSnapshot) => s.document.pages[s.appState.currentPageId].id
+const currentPageSelector = (s: TDSnapshot) => s.document.pages[s.appState.currentPageId]
 
 export function PageMenu() {
   const app = useTldrawApp()
@@ -43,7 +42,7 @@ export function PageMenu() {
     },
     [setIsOpen]
   )
-  const currentPageName = app.useStore(currentPageNameSelector)
+  const currentPageName = app.useStore(currentPageSelector).name
 
   return (
     <DropdownMenu.Root dir="ltr" open={isOpen} onOpenChange={handleOpenChange}>
@@ -62,7 +61,9 @@ function PageMenuContent({ onClose }: { onClose: () => void }) {
 
   const sortedPages = app.useStore(sortedSelector)
 
-  const currentPageId = app.useStore(currentPageIdSelector)
+  const currentPage = app.useStore(currentPageSelector)
+
+  const [gridType, setGridType] = React.useState(currentPage.gridType || 'none')
 
   const handleCreatePage = React.useCallback(() => {
     app.createPage()
@@ -76,9 +77,35 @@ function PageMenuContent({ onClose }: { onClose: () => void }) {
     [app]
   )
 
+  const handleGridTypeChange = React.useCallback((v) => {
+    setGridType(v)
+    app.setGridType(currentPage.id, v == 'none' ? undefined : v);
+  }, [app, setGridType, currentPage]);
+
   return (
     <>
-      <DropdownMenu.RadioGroup dir="ltr" value={currentPageId} onValueChange={handleChangePage}>
+      <DropdownMenu.RadioGroup dir="ltr" value={currentPage.id} onValueChange={handleChangePage}>
+        <ToggleGroup.Root type="single" value={gridType} onValueChange={handleGridTypeChange} orientation='horizontal'>
+          <ToggleGroup.Item value='none'>
+            <Cross2Icon/>
+          </ToggleGroup.Item>
+          <ToggleGroup.Item value='dots'>
+            <DragHandleDots2Icon />
+          </ToggleGroup.Item>
+          <ToggleGroup.Item value='squares'>
+            <FrameIcon />
+          </ToggleGroup.Item>
+          <ToggleGroup.Item value='lines'>
+            <TextAlignJustifyIcon />
+          </ToggleGroup.Item>
+          <ToggleGroup.Item value='iso'>
+            <TriangleRightIcon />
+          </ToggleGroup.Item>
+          <ToggleGroup.Item value='music'>
+            <SpeakerLoudIcon />
+          </ToggleGroup.Item>
+        </ToggleGroup.Root>
+        <DMDivider />
         {sortedPages.map((page) => (
           <ButtonWithOptions key={page.id}>
             <DropdownMenu.RadioItem
